third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-tables/html5-table-formatting-2-expected.txt

This is a testharness.js-based test.
FAIL Explicitely defined columns are not merged assert_equals: expected 100 but got 50
PASS Explicitely defined rows are not merged
FAIL Border-spacing is added between any two unmerged columns (1) assert_equals: expected 175 but got 100
PASS Border-spacing is added between any two unmerged rows (1)
PASS Border-spacing is added between any two unmerged columns (2)
PASS Border-spacing is added between any two unmerged rows (2)
PASS Border-spacing is added between any two unmerged columns (3)
PASS Border-spacing is added between any two unmerged rows (3)
PASS Border-spacing is added between any two unmerged columns (4)
PASS Border-spacing is added between any two unmerged rows (4)
FAIL Border-spacing is added between any two unmerged columns (5) assert_equals: expected 175 but got 100
PASS Border-spacing is added between any two unmerged rows (5)
Harness: the test ran to completion.
